Back End Development/Week 5 - Understanding Functions in Python/weather_api/weather_api/requirements.txt
attr==0.3.2
ConfigParser==7.0.0
contextlib2==21.6.0
cryptography==42.0.8
docutils==0.21.2
HTMLParser==0.0.2
importlib_metadata==8.0.0
ipython==8.12.3
ipywidgets==8.1.3
Jinja2==3.1.4
jnius==1.1.0
keyring==25.2.1
lockfile==0.12.2
mock==5.1.0
Pillow==10.4.0
protobuf==5.27.2
pyOpenSSL==24.1.0
python-dotenv==1.0.1
railroad==0.5.0
Sphinx==7.3.7
thread==2.0.3
tornado==6.4
truststore==0.9.1
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
